stock-a股股票基础研究/requirements.txt
tushare
pandas
python-dotenv
requests
pdfplumber
google-generativeai
